chore: move common hooks to peerprep/utils/client
