+++ apps/peerprep/src/layouts/auth-protected.tsx
@@ -10,12 +10,12 @@ import {
 } from "@peerprep/ui/dropdown-menu";
 import { Link } from "@peerprep/ui/link";
 import { getOrigin } from "@peerprep/utils/client";
+import { useAuth, useLogout } from "@peerprep/utils/client";
 import { ArrowUpRight } from "lucide-react";
 import toast from "react-hot-toast";
 import { Navigate, Outlet } from "react-router-dom";
 
 import { NavLogo } from "~/components/nav-logo";
-import { useAuth, useLogout } from "~/lib/auth";
 
 function NavAvatar() {
   const { data: user } = useAuth();
