+++ packages/utils/src/client/auth.ts
