+++ packages/utils/tsconfig.json
@@ -1,3 +1,6 @@
 {
-  "extends": "@peerprep/tsconfig/base.json"
+  "extends": "@peerprep/tsconfig/base.json",
+  "compilerOptions": {
+    "paths": { "~/*": ["./src/*"] }
+  }
 }
